Implemented visual territory generation on Android

+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/board/BoardJsonClasses.kt
@@ -18,6 +18,6 @@ data class TerritoryData(
 
 @Serializable
 data class XY(
-    val x: Float,
-    val y: Float
+    val x: Int,
+    val y: Int
 )
