+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/dataclasses/TerritoryRecord.kt
@@ -6,8 +6,8 @@ data class TerritoryRecord (
     val id: Int,
     var stat: Int,
     var continent: Continent,
-    val position: Pair<Float, Float>,
-    val size: Pair<Float, Float>,
+    val position: Pair<Int, Int>,
+    val size: Pair<Int, Int>,
 ) {
     val connections: MutableList<TerritoryRecord> = ArrayList()
     var owner: PlayerRecord? = null
